backend-python/requirements.txt
selenium>=4.16.0
undetected-chromedriver>=3.5.4
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
webdriver-manager>=4.0.1
fake-useragent>=1.4.0
